Refactor weapon input and shared weapon system
Move client input/update logic into a separate CEClientWeaponSystem.Input.cs partial file and adjust ClientUseItem signature and call sites. Add server-side handlers for client attack/stop requests (OnClientAttackRequest, OnClientStopRequest) and rename OnGetWeapon to OnGetWeaponAnimation. Clean up dependencies/usings, remove unused AdminLogger/PopupSystem, make CESharedAnimationActionSystem private (_animationAction) and update its usages, and remove the redundant CanAttack wrapper. General cleanup to separate input concerns and centralize server validation/animation handling.
